backend/requirements.txt
# Django & DRF
Django>=5.2.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0
django-extensions>=3.2.0
django-timezone-field>=6.0
drf-spectacular>=0.27.0

# Environment & Config
python-dotenv>=1.0.0
python-decouple>=3.8

# HTTP & API
requests>=2.31.0
httpx>=0.27.0

# AI & ML
google-genai>=1.0.0
google-generativeai>=0.8.0
numpy>=1.26.0
scipy>=1.11.0

# Firebase & Firestore
firebase-admin>=6.5.0

# Task Queue & Scheduling
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
python-crontab>=3.0.0
cron-descriptor>=1.4.0
gunicorn>=23.0.0
